refactor: App에서 받아오던 정보, 각 페이지마다 받아오도록 수정
