+++ src/pages/Main.jsx
@@ -1,7 +1,10 @@
-import React from 'react';
 import ProductListContainer from 'components/ProductList/ProductListContainer';
 
+import useLoadProducts from 'hooks/useLoadProducts';
+
 function Main() {
+  useLoadProducts();
+
   return <ProductListContainer />;
 }
